Requirement #1575: Watermarks for filesystem on Web UI should show absolute storage amount
This is update for this requirement. All the UI changes in previous commits have no effect as the decimal point precision in the database is not sufficient. We need to increase the precision to 4 decimal numbers (eg xx.yyyy). This will allow us to have 0.0001% accuracy for the high and low watermarks. For 1TB filesystem this translate to 1MB accuracy.

Changes in this commit:

- Increased the precision for HighWatermark and LowWatermark to 4 decimal points.

- Removed "PercentFull" field in Filesystem. This field is only used in the UI only as an informational field. This field was updated periodically but not used until users configure the filesystem via the web UI. We can obtain the live info instead by using the existing web service. It is also more preferable to use the "live" data especially when the users change the path, the UI should reflect the usage on the new path immediately.

- Change the javascript validation code to treat +ve error code as warning (ie. prompt users to decide whether to ignore the warning). When the wcf service is not available, a positive error code will be returned so that users can skip path validation.

+++ ImageServer/Web/Common/WebControls/RangeComparisonValidator.js
@@ -40,7 +40,7 @@ function @@CLIENTID@@_ClientSideEvaluator()
     
     if (this.input.value!=null && this.input.value!='' && !isNaN(this.input.value))
     {
-        controlValue = parseInt(this.input.value);
+        controlValue = parseFloat(this.input.value);
             
         if (compareCtrl!=null && compareCtrl.value!='' &&  !isNaN(compareCtrl.value))
         {
